refactor: add services layer
